test: add databaase event test (#2728)
* test: add tests and fix modify primary field bug

* test: add more test

* fix: tauri buiuld

* chore: disable share link button

+++ frontend/rust-lib/flowy-database2/src/services/database/database_editor.rs
@@ -10,7 +10,7 @@ use collab_database::views::{DatabaseLayout, DatabaseView, LayoutSetting};
 use parking_lot::Mutex;
 use tokio::sync::{broadcast, RwLock};
 
-use flowy_error::{internal_error, FlowyError, FlowyResult};
+use flowy_error::{internal_error, ErrorCode, FlowyError, FlowyResult};
 use flowy_task::TaskDispatcher;
 use lib_infra::future::{to_fut, Fut};
 
@@ -199,32 +199,32 @@ impl DatabaseEditor {
     }
   }
 
+  /// Returns a list of fields of the view.
+  /// If `field_ids` is not provided, all the fields will be returned in the order of the field that
+  /// defined in the view. Otherwise, the fields will be returned in the order of the `field_ids`.
   pub fn get_fields(&self, view_id: &str, field_ids: Option<Vec<String>>) -> Vec<Field> {
-    self.database.lock().get_fields(view_id, field_ids)
+    let database = self.database.lock();
+    let field_ids = field_ids.unwrap_or_else(|| {
+      database
+        .fields
+        .get_all_field_orders()
+        .into_iter()
+        .map(|field| field.id)
+        .collect()
+    });
+    database.get_fields(view_id, Some(field_ids))
   }
 
   pub async fn update_field(&self, params: FieldChangesetParams) -> FlowyResult<()> {
-    let is_primary = self
-      .database
-      .lock()
-      .fields
-      .get_field(&params.field_id)
-      .map(|field| field.is_primary)
-      .unwrap_or(false);
     self
       .database
       .lock()
       .fields
-      .update_field(&params.field_id, |mut update| {
-        update = update
+      .update_field(&params.field_id, |update| {
+        update
           .set_name_if_not_none(params.name)
           .set_width_at_if_not_none(params.width.map(|value| value as i64))
           .set_visibility_if_not_none(params.visibility);
-        if is_primary {
-          tracing::warn!("Cannot update primary field type");
-        } else {
-          update.set_field_type_if_not_none(params.field_type.map(|field_type| field_type.into()));
-        }
       });
     self
       .notify_did_update_database_field(&params.field_id)
@@ -233,6 +233,21 @@ impl DatabaseEditor {
   }
 
   pub async fn delete_field(&self, field_id: &str) -> FlowyResult<()> {
+    let is_primary = self
+      .database
+      .lock()
+      .fields
+      .get_field(field_id)
+      .map(|field| field.is_primary)
+      .unwrap_or(false);
+
+    if is_primary {
+      return Err(FlowyError::new(
+        ErrorCode::Internal,
+        "Can not delete primary field",
+      ));
+    }
+
     let database_id = {
       let database = self.database.lock();
       database.delete_field(field_id);
@@ -283,6 +298,13 @@ impl DatabaseEditor {
     match field {
       None => {},
       Some(field) => {
+        if field.is_primary {
+          return Err(FlowyError::new(
+            ErrorCode::Internal,
+            "Can not update primary field's field type",
+          ));
+        }
+
         let old_field_type = FieldType::from(field.field_type);
         let old_type_option = field.get_any_type_option(old_field_type.clone());
         let new_type_option = field
@@ -312,6 +334,21 @@ impl DatabaseEditor {
   }
 
   pub async fn duplicate_field(&self, view_id: &str, field_id: &str) -> FlowyResult<()> {
+    let is_primary = self
+      .database
+      .lock()
+      .fields
+      .get_field(field_id)
+      .map(|field| field.is_primary)
+      .unwrap_or(false);
+
+    if is_primary {
+      return Err(FlowyError::new(
+        ErrorCode::Internal,
+        "Can not duplicate primary field",
+      ));
+    }
+
     let value = self
       .database
       .lock()
